# PackageFluxApp-admin/docs/swagger.yaml
basePath: /api
definitions:
  domain.Detail:
    properties:
      appName:
        description: 应用名称
        type: string
      appleId:
        description: appleId
        type: string
      belongDept:
        description: 数据归属部门
        type: integer
      createTime:
        description: 创建时间
        type: string
      creator:
        description: 创建人
        type: integer
      description:
        description: 应用描述
        type: string
      id:
        description: 主键ID
        type: integer
      logoUrl:
        description: logo地址
        type: string
      modifier:
        description: 修改人
        type: integer
      preview:
        description: 预览图
        items:
          type: string
        type: array
      recordId:
        description: 记录ID
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示排序
        type: integer
      status:
        description: 状态
        type: boolean
      type:
        description: 1:APP
        type: string
      updateTime:
        description: 更新时间
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 用户
      user_id:
        description: 用户ID
        type: integer
      version:
        allOf:
        - $ref: '#/definitions/model.Version'
        description: 版本
      version_id:
        description: 版本ID
        type: integer
    type: object
...
  model.Detail:
    properties:
      appName:
        description: 应用名称
        type: string
      appleId:
        description: appleId
        type: string
      belongDept:
        description: 数据归属部门
        type: integer
      createTime:
        description: 创建时间
        type: string
      creator:
        description: 创建人
        type: integer
      description:
        description: 应用描述
        type: string
      id:
        description: 主键ID
        type: integer
      logoUrl:
        description: logo地址
        type: string
      modifier:
        description: 修改人
        type: integer
      preview:
        description: 预览图
        type: string
      recordId:
        description: 记录ID
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示排序
        type: integer
      status:
        description: 状态
        type: boolean
      type:
        description: 1:APP
        type: string
      updateTime:
        description: 修改时间
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.User'
        description: 用户
      user_id:
        description: 用户ID
        type: integer
      version:
        allOf:
        - $ref: '#/definitions/model.Version'
        description: 版本
      version_id:
        description: 版本ID
        type: integer
    type: object
  model.User:
    properties:
      avatar:
        description: 头像
        type: string
      belongDept:
        description: 数据归属部门
        type: integer
      createTime:
        description: 创建时间
        type: string
      creator:
        description: 创建人
        type: integer
      id:
        description: 主键ID
        type: integer
      modifier:
        description: 修改人
        type: integer
      recordId:
        description: 记录ID
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示排序
        type: integer
      status:
        description: 状态
        type: boolean
      total:
        description: 接口总次数
        type: integer
      updateTime:
        description: 修改时间
        type: string
      usageNumber:
        description: 接口调用次数
        type: integer
      username:
        description: 用户名
        type: string
    type: object
  model.Version:
    properties:
      apk_url:
        description: apk下载地址
        type: string
      belongDept:
        description: 数据归属部门
        type: integer
      createTime:
        description: 创建时间
        type: string
      creator:
        description: 创建人
        type: integer
      description:
        description: 更新说明
        type: string
      detail:
        allOf:
        - $ref: '#/definitions/model.Detail'
        description: 应用详情
      detail_id:
        description: 应用详情ID
        type: integer
      force_update:
        description: 是否强制更新
        type: boolean
      formal_version:
        description: 是否是正式版本
        type: boolean
      h_builder_update:
        description: HBuilderX是否更新
        type: boolean
      id:
        description: 主键ID
        type: integer
      modifier:
        description: 修改人
        type: integer
      recordId:
        description: 记录ID
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 显示排序
        type: integer
      status:
        description: 状态
        type: boolean
      updateTime:
        description: 修改时间
        type: string
      versionCode:
        description: 版本号
        type: string
      wgt_url:
        description: wgt下载地址
        type: string
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 提示信息
      success:
        description: 是否成功
        type: boolean
    type: object
  web.AppDetailsRequest:
    properties:
      appName:
        description: 应用名称
        maxLength: 50
        type: string
      appleId:
        description: appleId
        maxLength: 50
        type: string
      description:
        description: 应用描述
        maxLength: 255
        type: string
      logoUrl:
        description: logo地址
        type: string
      preview:
        description: 预览图
        items:
          type: string
        type: array
      remark:
        description: 备注
        maxLength: 255
        type: string
    required:
    - appName
    type: object
  web.DetailsListResponse:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/domain.Detail'
        type: array
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: 2224693191@qq.com
    name: CJiaの用心
    url: http://www.swagger.io/support
  description: PackageFluxApp-admin在线接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms
  title: PackageFluxApp-admin
  version: "1.0"
paths:
  /v1/application/details/create:
    post:
      consumes:
      - application/json
      description: 新增应用详情
      operationId: 新增应用详情
      parameters:
      - description: 应用详情参数
        in: body
        name: AppDetailsRequest
        required: true
        schema:
          $ref: '#/definitions/web.AppDetailsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增应用详情
      tags:
      - 应用详情
  /v1/application/details/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除应用详情
      operationId: 删除应用详情
      parameters:
      - description: 详情ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除应用详情
      tags:
      - 应用详情
  /v1/application/details/getById/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID查询应用详情
      operationId: 根据ID查询应用详情
      parameters:
      - description: 详情ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Detail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 根据ID查询应用详情
      tags:
      - 应用详情
  /v1/application/details/listAll:
    get:
      consumes:
      - application/json
      description: 查询应用详情列表
      operationId: 查询应用详情列表
      parameters:
      - default: "true"
        description: 状态
        in: query
        name: status
        required: true
        type: string
      - default: '""'
        description: 应用名称
        in: query
        name: appName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.DetailsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 查询应用详情列表
      tags:
      - 应用详情
  /v1/application/details/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新应用详情
      operationId: 更新应用详情
      parameters:
      - description: 详情ID
        in: path
        name: id
        required: true
        type: string
      - description: 应用详情参数
        in: body
        name: AppDetailsRequest
        required: true
        schema:
          $ref: '#/definitions/web.AppDetailsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新应用详情
      tags:
      - 应用详情
  /v1/application/files/batchUpload:
    post:
      consumes:
      - multipart/form-data
      description: 批量上传文件
      operationId: 批量上传文件
      parameters:
      - collectionFormat: csv
        description: 文件列表
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 批量上传文件
      tags:
      - 文件
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"